Implement single button functionality to control ACC time gap
This is a complete rewrite of the existing code. Use an Arduino with a CAN shield to send messages to change the car's adaptive cruise control (ACC) time gap setting. Send a 0x00 over the CAN, then alternate to a 0x01 when the button is pressed, and visa versa. Light up an LED hooked to the Arduino when the button is pressed.